Feature cordio update 20 05

+++ features/FEATURE_BLE/targets/TARGET_CORDIO_LL/stack/controller/sources/ble/bb/bb_ble_periodiclist.c
@@ -1,23 +1,24 @@
-/* Copyright (c) 2019 Arm Limited
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 /*************************************************************************************************/
 /*!
- * \file
- * \brief Generic BLE whitelist implementation file.
+ *  \file
+ *
+ *  \brief      Generic BLE whitelist implementation file.
+ *
+ *  Copyright (c) 2013-2018 Arm Ltd. All Rights Reserved.
+ *
+ *  Copyright (c) 2019-2020 Packetcraft, Inc.
+ *  
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 /*************************************************************************************************/
 
@@ -102,7 +103,7 @@ uint16_t BbBleInitPeriodicList(uint8_t numEntries, uint8_t *pFreeMem, uint32_t f
 /*************************************************************************************************/
 bool_t BbBlePeriodicListCheckAddr(uint8_t addrType, uint64_t addr, uint8_t SID)
 {
-  for (uint8_t i = 0; i < bbBlePeriodicListNumEntries; i++)
+  for (unsigned int i = 0; i < bbBlePeriodicListNumEntries; i++)
   {
     if ((pBbBlePeriodicListFilt[i].addrType == addrType) &&
         (pBbBlePeriodicListFilt[i].addr == addr) &&
@@ -134,8 +135,6 @@ uint8_t BbBlePeriodicListGetSize(void)
 /*************************************************************************************************/
 /*!
  *  \brief      Clear all periodic list entries.
- *
- *  \return     None.
  *
  *  Clear all periodic list entries stored in the BB.
  *
